pyproject.toml
[tool.poetry]
name = "peer-x"
version = "0.1.0"
description = ""
authors = ["gerome.vivar <gerome.vivar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
omegaconf = "^2.2.3"
hydra-core = "^1.2.0"
pytorch-lightning = "^1.7.6"

hydra-colorlog = "^1.2.0"
jupyterlab = "^3.4.7"
scikit-image = "^0.19.3"
torch = "^1.12.1"
torchvision = "^0.13.1"
matplotlib = "^3.6.0"
napari = "^0.4.16"
PyQt5 = "^5.15.7"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
